Montana/requirements.txt
# Montana Dependencies
# Install with: pip install -r requirements.txt

# Core cryptography
liboqs-python>=0.9.0     # Post-quantum cryptography (ML-DSA, ML-KEM)
cryptography>=41.0.0     # General cryptographic primitives

# Networking
aiohttp>=3.9.0           # Async HTTP client/server
websockets>=12.0         # WebSocket support

# Storage
lmdb>=1.4.0              # Lightning Memory-Mapped Database

# Optional: Telegram bot
python-telegram-bot>=20.0  # Telegram integration

# Development dependencies (install with: pip install -r requirements-dev.txt)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.0.0
# mypy>=1.0.0
# ruff>=0.1.0
# black>=23.0.0
